+++ lib/domain/model/entities.dart
@@ -717,7 +717,6 @@ class Entities<E extends ConceptEntity<E>> implements EntitiesApi<E> {
     }
     if (remove(beforeEntity)) {
       if (add(afterEntity)) {
-        afterEntity._whenUpdated = new DateTime.now();
         return true;
       } else {
         print('entities.update: ${errors.toList()}');